+++ CHANGELOG.md
@@ -3,6 +3,10 @@ SPDX-FileCopyrightText: 2025 Antoine Belvire
 SPDX-License-Identifier: GPL-3.0-or-later
 -->
 
+## Not released
+
+- GUI: Add dark theme support for puzzle thumbnail (!270). 
+
 ## Version 0.12 - Upgrade to Java 25 - 2025/11/09
 
 - Upgrade to Java 25 (!260).